AI/requirements-server.txt
flask
flask-socketio
python-dotenv
PyJWT
cryptography
flask-cors
flask-limiter
pytest
eventlet
pycryptodome
pydub
transformers